assetplan-mcp-agent/requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.3
pydantic~=2.11.7
pydantic-settings~=2.9.1
PyYAML~=6.0.2

langchain~=0.3.26
langchain-community~=0.3.26
langchain-core~=0.3.67
langgraph~=0.5.0
langgraph-checkpoint-postgres~=2.0.21
langchain-openai~=0.3.23
langchain-google-genai~=2.1.6
langchain-mcp-adapters==0.1.7
httpx